move env to dedicated files
